python-services/classifier/requirements.txt

# python-services/classifier/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-genai
python-dotenv==1.0.0